functions custom_layout and connect_ncbitaxa moved to new script: ete_layout.py
